Put DynamoAttributes to ZuoraAttributes converter within the comparison function. That's the only context in which it makes sense.
